# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Dokumentasi API untuk CRUD Admin WA Rotator",
        "version": "1.0"
    },
    "paths": {
...
        "/api/admin-admin": {
            "get": {
                "tags": [
                    "Admins"
                ],
                "summary": "Menampilkan daftar pengguna",
                "description": "Mengambil daftar pengguna dengan fitur pencarian berdasarkan username, email, atau role",
                "operationId": "e192d769ea9e1b70795e32eb7059269e",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Filter Admin berdasarkan username atau nomor hp",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Akmal"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Admin retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Users retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "username": {
                                                        "type": "string",
                                                        "example": "john_doe"
                                                    },
                                                    "nomor": {
                                                        "type": "string",
                                                        "example": "6281357477967"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/nomor-create": {
            "post": {
                "tags": [
                    "Admins"
                ],
                "summary": "Create a new nomor",
                "operationId": "a2b2936ce37dddf11866a14ce0999dec",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "nomor"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "johndoe"
                                    },
                                    "nomor": {
                                        "type": "string",
                                        "example": "6281357477967"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Nomor created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/nomor-update/{id}": {
            "put": {
                "tags": [
                    "Admins"
                ],
                "summary": "Update a nomor",
                "operationId": "99ba04efe1dc89e9972e134eadc92aa4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Nomor ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "nomor"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "john_updated"
                                    },
                                    "nomor": {
                                        "type": "string",
                                        "example": "6281357477999"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nomor updated successfully"
                    },
                    "404": {
                        "description": "Nomor not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/nomor-delete/{id}": {
            "delete": {
                "tags": [
                    "Admins"
                ],
                "summary": "Delete a nomor by ID",
                "operationId": "39ac09bcc7262727af0508979b2b26e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Nomor ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nomor deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Nomor deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nomor not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Nomor not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/show-link/{id?}": {
            "get": {
                "tags": [
                    "Admins"
                ],
                "summary": "Generate and show a link",
                "operationId": "f17f723bc36e5da29bf130284439b170",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Optional ID for generating a link",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Link generated successfully"
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/chat-show/{id}": {
            "get": {
                "tags": [
                    "Property"
                ],
                "summary": "Generate a chat message for a property",
                "operationId": "a5f3ea84cd5ed266c9b7d0807f20dc8e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Property ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chat message generated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Chat message generated successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "🌟 Halo Admin Opendoorz... (chat content)"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Property ID is required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Property ID is required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Property not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Property not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}